refactor(atomic): rename atomic_transaction_composer module to atomic, split into submodules
BREAKING CHANGE: the public module path algonaut::atomic_transaction_composer
is renamed to algonaut::atomic. No type names, signatures, or behaviour
change — the directory was named after the AtomicTransactionComposer type
that the typestate refactor (#307) deleted.

The 1115-line mod.rs is split into focused submodules: group (the public
typestate chain), method_call, encode, outcome, and signing. Submodules stay
private; the flat public re-exports are unchanged, so only the import path
changes.

Records the decision in docs/adr/atomic-module-layout.md and the breaking
import-path change in the changelog.
